Fixes the alignment of the tag svg for posts (#58)
When a post has tags, I think there was some unexpected markup on the page. The design mockup had just shown the tag names, but when tags are added to a post, the word "Tagged" is added above the tag names so that will always be at least two lines high.

This switches the alignment from the `bottom` to 3px from the top.