Adjusts the post footer meta layout
The post footer meta layout had been in columns that looked strange when content wrapped to another line.

Switched from `flexbox` to `grid` here to accomodate the layout for all users. Logged in users get the `.edit-link` element that should align to the top right of the `.entry-footer` container.

Also added icons for each piece of post footer meta. `tags` had an existing icon, but `categories` and `comments` needed icons.